[#660] Consolidate challenge methods into single module
* Merged challenge methods from `class_methods` into `challenges` to
simplify code.
* Removed the now redundant `class_methods` file and corresponding
`require`.
* Ensures all authenticity challenge logic is centralized and easily
maintained.
